requirements.txt
altair==5.5.0
asgiref==3.10.0
asttokens==3.0.0
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
bs4==0.0.2
cachetools==6.1.0
certifi==2025.7.9
cffi==1.17.1
charset-normalizer==3.4.2
chromedriver-binary==141.0.7354.0.0
click==8.2.1
colorama==0.4.6
comm==0.2.2
contourpy==1.3.2
cycler==0.12.1
debugpy==1.8.14
decorator==5.2.1
Django==5.2.8
et_xmlfile==2.0.0
executing==2.2.0
fonttools==4.58.1
gitdb==4.0.12
GitPython==3.1.44
h11==0.16.0
idna==3.10
ipykernel==6.29.5
ipython==9.4.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
narwhals==1.48.0
nest-asyncio==1.6.0
numpy==2.2.6
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.0
parso==0.8.4
pillow==11.2.1
platformdirs==4.3.8
prompt_toolkit==3.0.51
protobuf==6.31.1
psutil==7.0.0
pure_eval==0.2.3
pyarrow==21.0.0
pycparser==2.22
pydeck==0.9.1
Pygments==2.19.2
pyparsing==3.2.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2025.2
pywin32==310
pyzmq==27.0.0
referencing==0.36.2
requests==2.32.4
rpds-py==0.26.0
seaborn==0.13.2
selenium==4.35.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
sqlparse==0.5.3
stack-data==0.6.3
streamlit==1.48.0
tenacity==9.1.2
toml==0.10.2
tornado==6.5.1
traitlets==5.14.3
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
watchdog==6.0.0
wcwidth==0.2.13
websocket-client==1.8.0
wsproto==1.2.0
